+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fix obsolete subcommands in the documentation
 * Add a timeout option when using `deploy` subcommand (default is 1440 minutes)
+* New args for localtest: `--no-tests`
 
 ## \[1.0.0\] - 2023-07-18